feat(domain service controller mapper ): 增加使用德鲁伊数据源，修复一些bug
修复单位问题

[skip ci]

+++ src/main/java/com/feidian/ChromosView/domain/MatrixPoint.java
@@ -11,8 +11,8 @@ import java.io.Serializable;
 @AllArgsConstructor
 @NoArgsConstructor
 public class MatrixPoint implements Serializable {
-    long binX;
-    long binY;
+    double binX;
+    double binY;
     long genomeX;
     long genomeY;
     float counts;
